requirements.txt
typing-extensions
langchain-openai
langchain-experimental
langchain_anthropic
langchain_google_genai
pandas
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
chainlit
rich
questionary
ollama
pytest
python-dotenv
